+++ SDLAssignment/Vector2D.h
@@ -5,7 +5,7 @@ class Vector2D
 
 public:
 
-	Vector2D(double x = 0, double y = 0);
+	Vector2D(int x = 0, int y = 0);
 
 	Vector2D Add(const Vector2D& secondVector);
 	Vector2D Subtract(const Vector2D& secondVector);
@@ -14,11 +14,11 @@ public:
 	Vector2D Negate();
 	Vector2D Normalised();
 
-	double Distance(const Vector2D& secondVector);
-	double Magnitude();
+	int Distance(const Vector2D& secondVector);
+	int Magnitude();
 
-	double x;
-	double y;
+	int x;
+	int y;
 
 private:
